+++ src/hump_yard/__init__.py
@@ -4,4 +4,4 @@ from .base_plugin import FileProcessorPlugin, PluginManager
 
 __all__ = ['FileMonitorDaemon', 'FileProcessorPlugin', 'PluginManager']
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'